Fix up more types and cleanup old code
